refactor: move protocol negotiation to client specific
